frontend/src/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
groq==0.9.0
python-dotenv==1.0.0
websockets==13.0
requests==2.32.0
pyautogui==0.9.54
pygetwindow==0.0.9
pyperclip==1.8.2
